VoiceScribeWindows/requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.5.0
numpy>=1.24.0
pystray>=0.19.0
Pillow>=10.0.0
keyboard>=0.13.5
pyperclip>=1.8.0
pyautogui>=0.9.54
